+++ src/buf.c
@@ -53,19 +53,6 @@
 #include "job.h"
 #include "util.h"
 
-/* use fseeko instead of fseek for long file support if we have it */
-#ifdef HAVE_FSEEKO
-#define fseek fseeko
-#elif defined HAVE_FSEEKO64
-#define fseek fseeko64
-#endif
-
-/**
- * File IO buffer sizes.
- */
-int rs_inbuflen = 16000, rs_outbuflen = 16000;
-
-
 struct rs_filebuf {
         FILE *f;
         char            *buf;
@@ -101,10 +88,10 @@ void rs_filebuf_free(rs_filebuf_t *fb)
 rs_result rs_infilebuf_fill(rs_job_t *job, rs_buffers_t *buf,
                             void *opaque)
 {
-    int                     len;
+    size_t                  len;
     rs_filebuf_t            *fb = (rs_filebuf_t *) opaque;
     FILE                    *f = fb->f;
-        
+
     /* This is only allowed if either the buf has no input buffer
      * yet, or that buffer could possibly be BUF. */
     if (buf->next_in != NULL) {
@@ -125,23 +112,20 @@ rs_result rs_infilebuf_fill(rs_job_t *job, rs_buffers_t *buf,
         /* Still some data remaining.  Perhaps we should read
            anyhow? */
         return RS_DONE;
-        
+
     len = fread(fb->buf, 1, fb->buf_len, f);
     if (len <= 0) {
-        /* This will happen if file size is a multiple of input block len
-         */
+        /* This will happen if file size is a multiple of input block len */
         if (feof(f)) {
             rs_trace("seen end of file on input");
             buf->eof_in = 1;
             return RS_DONE;
         }
         if (ferror(f)) {
-            rs_error("error filling buf from file: %s",
-                     strerror(errno));
+            rs_error("error filling buf from file: %s", strerror(errno));
             return RS_IO_ERROR;
         } else {
-            rs_error("no error bit, but got %d return when trying to read",
-                     len);
+            rs_error("no error bit, but got "FMT_SIZE" return when trying to read", len);
             return RS_IO_ERROR;
         }
     }
@@ -169,13 +153,13 @@ rs_result rs_outfilebuf_drain(rs_job_t *job, rs_buffers_t *buf, void *opaque)
      * yet, or that buffer could possibly be BUF. */
     if (buf->next_out == NULL) {
         assert(buf->avail_out == 0);
-                
+
         buf->next_out = fb->buf;
         buf->avail_out = fb->buf_len;
-                
+
         return RS_DONE;
     }
-        
+
     assert(buf->avail_out <= fb->buf_len);
     assert(buf->next_out >= fb->buf);
     assert(buf->next_out <= fb->buf + fb->buf_len);
@@ -183,13 +167,12 @@ rs_result rs_outfilebuf_drain(rs_job_t *job, rs_buffers_t *buf, void *opaque)
     present = buf->next_out - fb->buf;
     if (present > 0) {
         int result;
-                
+
         assert(present > 0);
 
         result = fwrite(fb->buf, 1, present, f);
         if (present != result) {
-            rs_error("error draining buf to file: %s",
-                     strerror(errno));
+            rs_error("error draining buf to file: %s", strerror(errno));
             return RS_IO_ERROR;
         }
 
@@ -198,30 +181,6 @@ rs_result rs_outfilebuf_drain(rs_job_t *job, rs_buffers_t *buf, void *opaque)
 
         job->stats.out_bytes += result;
     }
-        
+
     return RS_DONE;
-}
-
-
-rs_result rs_file_copy_cb(void *arg, rs_long_t pos, size_t *len, void **buf)
-{
-    int        got;
-    FILE       *f = (FILE *) arg;
-
-    if (fseek(f, pos, SEEK_SET)) {
-        rs_log(RS_LOG_ERR, "seek failed: %s", strerror(errno));
-        return RS_IO_ERROR;
-    }
-
-    got = fread(*buf, 1, *len, f);
-    if (got == -1) {
-        rs_error("read error: %s", strerror(errno));
-        return RS_IO_ERROR;
-    } else if (got == 0) {
-        rs_error("unexpected eof on fd%d", fileno(f));
-        return RS_INPUT_ENDED;
-    } else {
-        *len = got;
-        return RS_DONE;
-    }
 }