+++ src/util.c
@@ -26,12 +26,8 @@
 
 
 #include "config.h"
-#include <sys/types.h>
-
 #include <stdlib.h>
 #include <string.h>
-#include <stdio.h>
-#include <sys/stat.h>
 
 #include "librsync.h"
 #include "util.h"
@@ -71,31 +67,13 @@ rs_alloc(size_t size, char const *name)
 }
 
 
-#ifdef HAVE_FSTATI64
-#  ifdef stat
-#   undef stat
-#  endif
-#  define stat _stati64
-#  ifdef fstat
-#   undef fstat
-#  endif
-#  define fstat(f,s) _fstati64((f), (s))
-#elif defined HAVE_FSTAT64
-#  ifdef stat
-#   undef stat
-#  endif
-#  define stat stat64
-#  ifdef fstat
-#   undef fstat
-#  endif
-#  define fstat(f,s) fstat64((f), (s))
-#endif
-
-void
-rs_get_filesize(FILE *f, rs_long_t *size)
+void *
+rs_realloc(void *ptr, size_t size, char const *name)
 {
-    struct stat st;
-    if (size && (fstat(fileno(f), &st) == 0) && (S_ISREG(st.st_mode))) {
-        *size = st.st_size;
-    }
+    void *p;
+
+    if (!(p = realloc(ptr, size))) {
+	rs_fatal("couldn't reallocate instance of %s", name);
+     }
+     return p;
 }